added user mode checks

+++ bin/clean.sh
@@ -1,6 +1,13 @@
 #!/usr/bin/env bash
 set -e
 
+uid=$(id -u)
+
+if [ "${uid}" -eq 0 ]; then
+	echo "Please run as user"
+	exit
+fi
+
 pwd=$(pwd)
 
 git clean -f -q -d -x
